refactpred the audit event processing logic into it's own class to support better unit testing and organization
