+++ DotNet/Account/appsettings.Development.json
@@ -30,7 +30,7 @@
     "KeyRing": "Link"
   },
   "Authentication": {
-    "EnableAnonymousAccess": false,
+    "EnableAnonymousAccess": true,
     "Schemas": {
       "LinkBearer": {
         "Authority": "https://localhost:7004",
